# src/main/java/com/smsverification/gateway/api/ApiError.java
package com.smsverification.gateway.api;

import java.time.Instant;
import java.util.Map;

public record ApiError(
        Instant timestamp,
        int status,
        String error,
        String code,
        String message,
        String path,
        Map<String, String> fieldErrors
) {
}