# src/main/java/com/smsverification/gateway/sms/IncomingSmsResponse.java
package com.smsverification.gateway.sms;

import com.smsverification.gateway.verification.VerificationStatus;

import java.util.UUID;

public record IncomingSmsResponse(
        UUID smsId,
        boolean duplicate,
        IncomingSmsMatchStatus matchStatus,
        UUID verificationId,
        VerificationStatus verificationStatus
) {
}