# src/main/java/com/smsverification/gateway/verification/VerificationController.java
package com.smsverification.gateway.verification;

import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

import java.util.UUID;

@RestController
@RequestMapping("/api/v1/verifications")
public class VerificationController {

    private final VerificationService verificationService;

    public VerificationController(VerificationService verificationService) {
        this.verificationService = verificationService;
    }

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    public CreateVerificationResponse create(@Valid @RequestBody CreateVerificationRequest request) {
        return verificationService.create(request);
    }

    @GetMapping("/{verificationId}/status")
    public VerificationStatusResponse getStatus(@PathVariable UUID verificationId) {
        return verificationService.getStatus(verificationId);
    }
}